requirements.txt
meta-ads-mcp
facebook-business>=22.0.0
python-dotenv>=1.0.0
